refactor(notes): rename folder
