+++ src/modules/notes/NoteEdit/services/fetchGetOneNoteService.ts
@@ -1,4 +1,4 @@
-import { IFetchGetOneNoteResponse } from "modules/notes/NoteEdit/service/types";
+import { IFetchGetOneNoteResponse } from "modules/notes/NoteEdit/services/types";
 import api from "shared/services/api";
 
 export const fetchGetOneNoteService = async (noteID: string | null) => {
