+++ src/modules/notes/NoteEdit/services/fetchPostTodosService.ts
@@ -1,6 +1,6 @@
 import { IFetchResponseDefault } from "shared/common/types/Fetch";
 import api from "shared/services/api";
-import { IFetchPostTodosService } from "modules/notes/NoteEdit/service/types/fetchPostTodosService.types";
+import { IFetchPostTodosService } from "modules/notes/NoteEdit/services/types";
 
 export const fetchPostTodosService: IFetchPostTodosService = async ({
 	body,
