+++ src/modules/notes/NoteEdit/services/types/index.ts
